add validation input on backend

+++ server/controllers/postControllers.js
@@ -22,6 +22,18 @@ export const generatePostCaptions = async (req, res) => {
   try {
     const { socialNetwork, subject, tone } = req.body;
 
+    if (!socialNetwork) {
+      return res.status(400).send("Social network is required");
+    }
+
+    if (!subject) {
+      return res.status(400).send("Subject is required");
+    }
+
+    if (!tone) {
+      return res.status(400).send("Tone is required");
+    }
+
     const result = await generatePostCaptionsByAI(socialNetwork, subject, tone);
 
     res.status(200).send(result);
@@ -33,6 +45,10 @@ export const generatePostCaptions = async (req, res) => {
 export const getPostIdeas = async (req, res) => {
   try {
     const { topic } = req.body;
+
+    if (!topic) {
+      return res.status(400).send("Topic is required");
+    }
 
     const result = await getPostIdeasByAI(topic);
 
@@ -46,6 +62,10 @@ export const createCaptionsFromIdeas = async (req, res) => {
   try {
     const { idea } = req.body;
 
+    if (!idea) {
+      return res.status(400).send("Idea is required");
+    }
+
     const result = await generateCaptionsFromIdeasByAI(idea);
 
     res.status(200).send(result);
@@ -57,6 +77,18 @@ export const createCaptionsFromIdeas = async (req, res) => {
 export const saveGeneratedContent = async (req, res) => {
   try {
     const { topic, data, phoneNumber } = req.body;
+
+    if (!topic) {
+      return res.status(400).send("Topic is required");
+    }
+
+    if (!data) {
+      return res.status(400).send("Data is required");
+    }
+
+    if (!phoneNumber) {
+      return res.status(400).send("Phone number is required");
+    }
 
     const docRef = await addDoc(collection(db, "savedPosts"), {
       topic: topic.toLowerCase().trim(),
@@ -76,6 +108,10 @@ export const saveGeneratedContent = async (req, res) => {
 export const getUserGeneratedContents = async (req, res) => {
   try {
     const { phoneNumber } = req.params;
+
+    if (!phoneNumber) {
+      return res.status(400).send("Phone number is required");
+    }
 
     const q = query(
       collection(db, "savedPosts"),
@@ -101,6 +137,10 @@ export const getUserGeneratedContents = async (req, res) => {
 export const unsaveContent = async (req, res) => {
   try {
     const { captionId } = req.params;
+
+    if (!captionId) {
+      return res.status(400).send("Caption Id is required");
+    }
 
     await deleteDoc(doc(db, "savedPosts", captionId));
